Refactor ConfigWindowComponent
