+++ src/AzureOpenAIProxy.PlaygroundApp/Components/Pages/Playground.razor
@@ -8,8 +8,7 @@
 
     <FluentGrid Spacing="3" AdaptiveRendering="true" Justify="JustifyContent.FlexStart">
         <FluentGridItem Class="config-grid" xs="12" sm="12" md="4" lg="4" xl="4" xxl="4">
-            <ApiKeyInputComponent Id="api-key" @rendermode="InteractiveServer" />
-            <ConfigTabComponent @rendermode="InteractiveServer"/>
+            <ConfigWindowComponent Id="config-window" @rendermode="InteractiveServer" />
         </FluentGridItem>
 
         <FluentGridItem Class="chat-grid" xs="12" sm="12" md="8" lg="8" xl="8" xxl="8" Style="height: 900px;">
