+++ src/AzureOpenAIProxy.PlaygroundApp/Components/UI/ConfigTabComponent.razor
@@ -1,20 +1,23 @@
-<FluentTabs Id="config-tab" ActiveTabId="system-message" OnTabChange="ChangeTab">
-    <FluentTab Label="System message" Id="system-message-tab">
-        This is "System message" tab.
-    </FluentTab>
-    <FluentTab Label="Parameters" Id="parameters-tab">
-        This is "Parameters" tab.
-    </FluentTab>
-</FluentTabs>
-
-<p id="active-tab">[TEST] Active tab changed to: @SelectedTab?.Id</p>
+<FluentStack Id="@Id" Style="width: 100%;" Orientation="Orientation.Vertical" VerticalAlignment="VerticalAlignment.Top">
+    <FluentTabs Id="config-tabs" Style="width: 100%;" ActiveTabId="system-message-tab" OnTabChange="ChangeTab">
+        <FluentTab Id="system-message-tab" Label="System message">
+            This is "System message" tab.
+        </FluentTab>
+        <FluentTab Id="parameters-tab" Label="Parameters">
+            This is "Parameters" tab.
+        </FluentTab>
+    </FluentTabs>
+</FluentStack>
 
 @code {
-    FluentTab? SelectedTab;
+    private FluentTab? selectedTab { get; set; }
+
+    [Parameter]
+    public string? Id { get; set; }
 
     private async Task ChangeTab(FluentTab tab)
     {
-        SelectedTab = tab;
+        this.selectedTab = tab;
         await Task.CompletedTask;
     }
 }
