+++ src/AzureOpenAIProxy.PlaygroundApp/Components/UI/ConfigWindowComponent.razor
@@ -1,0 +1,31 @@
+﻿@using AzureOpenAIProxy.PlaygroundApp.Models
+
+<FluentStack Id="@Id" Style="width: 100%; height: 100%;" Orientation="Orientation.Vertical" VerticalAlignment="VerticalAlignment.Top">
+    <FluentLabel Typo="Typography.H2">Setup</FluentLabel>
+    <ApiKeyInputComponent Id="api-key" OnKeyInput="SetApiKey" @rendermode="InteractiveServer" />
+    <DeploymentModelListComponent Id="deployment-model-list" OnUserOptionSelected="SetDeploymentModel" @rendermode="InteractiveServer" />
+    <FluentDivider />
+    <ConfigTabComponent id="config-tab" @rendermode="InteractiveServer" />
+</FluentStack>
+
+@code {
+    private string? apiKey;
+    private string? deploymentModel;
+
+    [Parameter]
+    public string? Id { get; set; }
+
+    private async Task SetApiKey(string apiKey)
+    {
+        this.apiKey = apiKey;
+
+        await Task.CompletedTask;
+    }
+
+    private async Task SetDeploymentModel(string deploymentModel)
+    {
+        this.deploymentModel = deploymentModel;
+
+        await Task.CompletedTask;
+    }
+}
