--- src/AzureOpenAIProxy.PlaygroundApp/Components/UI/DebugButtonComponent.razor
@@ -1,25 +1,0 @@
-@inject IToastService ToastService
-
-<!-- This component displays the given input when the button is clicked. -->
-<FluentButton Id="@Id" @onclick="ShowToast" Appearance="Appearance.Accent">Debug</FluentButton>
-
-@code {
-    [Parameter]
-    public string? Id { get; set; }
-
-    [Parameter]
-    public object? Input { get; set; }
-    
-    private async Task ShowToast()
-    {
-        if (Input is null)
-        {
-            ToastService.ShowToast(ToastIntent.Warning, "Input is null.");
-            await Task.CompletedTask;
-            return;
-        }
-
-        ToastService.ShowToast(ToastIntent.Success, $"{Input} (Type: {Input.GetType()})");
-        await Task.CompletedTask;
-    }
-}
