--- src/AzureOpenAIProxy.PlaygroundApp/Components/UI/DebugTargetComponent.razor
@@ -1,18 +1,0 @@
-<FluentRadioGroup Id="@Id" Label="UI Test #171" TValue="int" ValueChanged="SetValue" Orientation="Orientation.Vertical">
-    <FluentRadio Value="123">123</FluentRadio>
-    <FluentRadio Value="456">456</FluentRadio>
-    <FluentRadio Value="789">789</FluentRadio>
-</FluentRadioGroup>
-
-@code {
-    [Parameter]
-    public string? Id { get; set; }
-
-    [Parameter]
-    public EventCallback<int> OnValueChanged { get; set; }
-
-    private async Task SetValue(int value)
-    {
-        await OnValueChanged.InvokeAsync(value);
-    }
-}
